[substrate_api] remove duplicated definition of reconnecting_ws_provider.dart
