+++ app/lib/service/substrate_api/api.dart
@@ -13,6 +13,7 @@ import 'package:encointer_wallet/service/substrate_api/chain_api.dart';
 import 'package:encointer_wallet/service/substrate_api/core/dart_api.dart';
 import 'package:encointer_wallet/service/substrate_api/core/js_api.dart';
 import 'package:encointer_wallet/service/substrate_api/encointer/encointer_api.dart';
+import 'package:encointer_wallet/service/substrate_api/core/reconnecting_ws_provider.dart';
 import 'package:ew_polkadart/ew_polkadart.dart';
 
 /// Global api instance
@@ -238,76 +239,5 @@ class Api {
   Future<void> closeWebView() async {
     await stopSubscriptions();
     return js.closeWebView();
-  }
-}
-
-class ReconnectingWsProvider extends Provider {
-  ReconnectingWsProvider(this.url, {bool autoConnect = true})
-      : provider = WsProvider(
-          url,
-          autoConnect: autoConnect,
-        );
-
-  final Uri url;
-  WsProvider provider;
-
-  Future<void> connectToNewEndpoint(Uri url) async {
-    await disconnect();
-    provider = WsProvider(url);
-  }
-
-  @override
-  Future connect() {
-    if (isConnected()) {
-      return Future.value();
-    } else {
-      // We want to use a new channel even if the channel exists but it was closed.
-      provider.channel = null;
-      provider = WsProvider(url, autoConnect: false);
-      return provider.connect();
-    }
-  }
-
-  @override
-  Future disconnect() async {
-    // We only care if the channel is not equal to null.
-    // Because we still want the internal cleanup if
-    // the connection was closed from the other end.
-    if (provider.channel == null) {
-      return Future.value();
-    } else {
-      try {
-        await provider.disconnect();
-      } catch (e) {
-        Log.e('Error disconnecting websocket: $e', 'ReconnectingWsProvider');
-        return Future.value();
-      }
-    }
-  }
-
-  @override
-  bool isConnected() {
-    // the `provider.isConnected()` check is wrong upstream.
-    // Hence, we implement it ourselves.
-    final channel = provider.channel;
-    return channel != null && channel.closeCode == null;
-  }
-
-  @override
-  Future<RpcResponse> send(String method, List<dynamic> params) async {
-    // Connect if disconnected
-    await connect();
-    return provider.send(method, params);
-  }
-
-  @override
-  Future<SubscriptionResponse> subscribe(
-    String method,
-    List<dynamic> params, {
-    FutureOr<void> Function(String subscription)? onCancel,
-  }) async {
-    // Connect if disconnected
-    await connect();
-    return provider.subscribe(method, params, onCancel: onCancel);
   }
 }
