+++ app/test/mock/api/mock_polkadart_provider.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:encointer_wallet/service/service.dart';
+import 'package:encointer_wallet/service/substrate_api/core/reconnecting_ws_provider.dart';
 import 'package:ew_polkadart/ew_polkadart.dart';
 
 class MockPolkadartProvider extends ReconnectingWsProvider {
